+++ app/models/pydantic/PermissionsModel.py
@@ -14,3 +14,16 @@ class PydanticPermissionsModel(BaseModel):
 
     class Config:
         from_attributes = True
+
+class PydanticPermissionModelFromJSON(BaseModel):
+    """
+    Pydantic model for permissions. Used to load data from JSON files.
+    """
+    service_name_id    : str
+    operation_name_id  : str
+
+    class Config:
+        """
+        Pydantic configuration for the model.
+        """
+        from_attributes = True
