+++ app/utils/databases/datasets.py
@@ -14,7 +14,7 @@ from app.models.pydantic.AccountModel import PydanticAccountModelFromJSON
 from app.models.pydantic.CoefficientModel import PydanticCoefficientModelFromJSON
 from app.models.pydantic.CourseTypeModel import PydanticCourseTypeModelFromJSON
 from app.models.pydantic.OperationModel import PydanticOperationModelFromJSON
-from app.models.pydantic.PermissionModel import PydanticPermissionModelFromJSON
+from app.models.pydantic.PermissionsModel import PydanticPermissionModelFromJSON
 from app.models.pydantic.ProfileModel import PydanticProfileModelFromJSON
 from app.models.pydantic.PydanticRole import PydanticRoleModelFromJSON
 from app.models.pydantic.ServiceModel import PydanticServiceModelFromJSON
